Added a message event handler base class and reaction event handler base class
Rational: To consolidate re-used methods (partial handlers and handleMessageCheck()).
