+++ src/main/eventhandler/MessageUpdateEventHandler.ts
@@ -1,30 +1,8 @@
-import { Message } from 'discord.js';
-import { EventHandler } from './EventHandler';
-import { Storage } from '../storage/Storage';
-import { MessageChecker } from '../modules/messagechecker/MessageChecker';
-import { MessageResponse } from '../modules/messagechecker/response/MessageResponse';
+import { MessageEventHandler } from './MessageEventHandler';
 
-export class MessageUpdateEventHandler extends EventHandler {
-    protected message: Message;
-
-    public constructor(storage: Storage, message: Message) {
-        super(storage);
-        this.message = message;
-    }
-
+export class MessageUpdateEventHandler extends MessageEventHandler {
     /**
-     * Handles fetching of message if it's partial.
-     *
-     * @returns Promise<void>
-     */
-    public async handlePartial(): Promise<void> {
-        if (this.message.partial) {
-            await this.message.fetch();
-        }
-    }
-
-    /**
-     * Handles when message is editted
+     * Handles when message is edited
      *
      * @returns Promise
      */
@@ -36,22 +14,8 @@ export class MessageUpdateEventHandler extends EventHandler {
         // If it's a bot, ignore :)
         if (this.message.author.bot) return;
 
-        const server = this.getServer(this.message.guild.id.toString());
-
-        // Retrieve settings
-        const bannedWords = server.messageCheckerSettings.getBannedWords();
-        const reportingChannelId = server.messageCheckerSettings.getReportingChannelId();
-        const responseMessage = server.messageCheckerSettings.getResponseMessage();
-        const deleteMessage = server.messageCheckerSettings.getDeleteMessage();
-
         // Check updated message
-        const result
-            = await new MessageChecker().checkMessage(this.message.content, bannedWords);
-        if (result.guilty) {
-            new MessageResponse(this.message)
-                .sendReport(result, reportingChannelId)
-                .sendMessageToUser(responseMessage)
-                .deleteMessage(deleteMessage);
-        }
+        const server = this.getServer(this.message.guild.id.toString());
+        this.handleMessageCheck(server);
     }
 }
